Split nvme to a file
